requirements.txt
# Following libraries are needed for the proyect 
pytest==8.1.1
pytest-cov==5.0.0
Flask==3.0.2
Flask-Testing==0.8.1
Flask-Cors==4.0.0
huggingface_hub==0.21.4
langchain==0.1.13
python-dotenv==1.0.1
openai==1.14.2
Werkzeug==3.0.1
langchain_openai==0.1.1 
regex==2023.12.25 
tiktoken==0.6.0
beautifulsoup4==4.9.3 
nltk == 3.6.2
graphviz==0.17
datetime==4.3

# pip install -r requirements.txt
